Refactor SpanWriter and SpanReader into separate files
